+++ lib/fsm_hit_detection/accelerometer.h
@@ -20,8 +20,8 @@ void accel_set_thresholds ( uint8_t index, accel_threshold_t hard, accel_thresho
 void accel_set_pins ( uint8_t index, uint8_t pin_SDA, uint8_t pin_SCL );
 int8_t accel_get_pins ( uint8_t index, uint8_t* pin_SDA, uint8_t* pin_SCL );
 
-void accel_set_timeout ( uint8_t index, uint16_t new_timeout);
-uint16_t accel_get_timeout (uint8_t index);
+void accel_set_timeout ( uint8_t index, uint32_t new_timeout);
+uint32_t accel_get_timeout (uint8_t index);
 
 void accel_set_active (uint8_t index, uint8_t val);
 uint8_t accel_is_active (uint8_t index);
@@ -32,8 +32,8 @@ uint8_t accel_get_flag_crash (uint8_t index);
 void accel_set_flag_movement (uint8_t index, uint8_t val);
 uint8_t accel_get_flag_movement (uint8_t index);
 
-void accel_set_sampling ( uint8_t index, uint16_t val);
-uint16_t accel_get_sampling ( uint16_t index );
+void accel_set_sampling ( uint8_t index, uint32_t val);
+uint32_t accel_get_sampling ( uint8_t index );
 
 void accel_set_accel_xyz (uint8_t index, uint16_t accel_x, uint16_t accel_y, uint16_t accel_z);
 void accel_get_accel_xyz (uint8_t index, uint16_t* accel_x, uint16_t* accel_y, uint16_t* accel_z);
